feat: add tarjeta credito

+++ src/Domain/Patterns/StateTarjetaCredito/IEstadoTarjetaCredito.cs
@@ -4,7 +4,7 @@ namespace Domain.Interfaces.States
 {
     public interface IEstadoTarjeta
     {
-        void RealizarCompra(TarjetaCredito tarjeta, decimal monto);
-        void PagarTarjeta(TarjetaCredito tarjeta, decimal monto);
+        void RealizarCompra(TarjetaCredito tarjeta, double monto);
+        void PagarTarjeta(TarjetaCredito tarjeta, double monto);
     }
 }
